refactor controllers -> added model layer. added missing types
