+++ src/controller/PatientController.ts
@@ -1,66 +1,18 @@
-import { appointments, patients } from "@prisma/client";
-import { PatientRequestType, TreatmentGroupType } from "~/types/patient";
-import database from "~/model/prisma";
-
-const findPatient = async function(id: bigint): Promise<patients | null> {
-    return await database.patients.findUnique({
-        where: {
-            id,
-        }
-    });
-}
-
+import { PatientRequestType, PatientResponseType, TreatmentGroupType } from "~/types/patient";
+import patients from "~/model/actions/Patients";
+import { AppointmentResponseType } from "~/types/appointment";
 class PatientController {
-    async getProfile(patient: PatientRequestType): Promise<patients> {
-        const patient_model = await database.patients.findUnique({
-            where: {
-                id: patient.patient_id
-            }
-        });
-
-        if (patient_model == null) {
-            throw new Error(`Patient with id ${patient.patient_id} does not exist.`);
-        }
-
-        return patient_model;
+    async getProfile(patient: PatientRequestType): Promise<PatientResponseType> {
+        return await patients.getProfile(patient);
     }
 
-    async getAppointments(patient: PatientRequestType): Promise<Array<appointments>> {
-        if (await findPatient(patient.patient_id) == null) {
-            throw new Error(`Patient with id ${patient.patient_id} does not exist.`);
-        }
-
-        const appointments = await database.appointments.findMany({
-            where: {
-                patient_id: patient.patient_id,
-            }
-        });
-
-        return appointments;
+    async getAppointments(patient: PatientRequestType): Promise<Array<AppointmentResponseType>> {
+        return await patients.getAppointments(patient);
     }
 
     async getTreatments(patient: PatientRequestType): Promise<Array<TreatmentGroupType>> {
-        if (await findPatient(patient.patient_id) == null) {
-            throw new Error(`Patient with id ${patient.patient_id} does not exist.`);
-        }
-
-        const treatments = await database.stories.findMany({
-            select: {
-                date: true,
-                treatments: {
-                    select: {
-                        title: true,
-                        description: true,
-                    }
-                }
-            },
-            where: {
-                patient_id: patient.patient_id,
-            }
-        });
-
-        return treatments;
-    }
+        return await patients.getTreatments(patient);
+    }   
 }
 
 export default new PatientController();
