+++ src/types/patient.ts
@@ -11,3 +11,12 @@ export type TreatmentGroupType = {
     date: Date,
     treatments: NarrowTreatmentType[],
 }
+
+export type PatientResponseType = {
+    id: bigint,
+    name: string,
+    lastname: string,
+    email: string,
+    health_insurance_company: string | null,
+    health_insurance_id: string | null,
+}
